Thêm mới user thành công

+++ src/main/java/vn/hunghaohan/service/UserService.java
@@ -1,18 +1,25 @@
 package vn.hunghaohan.service;
 
-import vn.hunghaohan.controller.request.user.UserCreationRequest;
-import vn.hunghaohan.controller.request.user.UserPasswordRequest;
-import vn.hunghaohan.controller.request.user.UserUpdateRequest;
+import vn.hunghaohan.controller.request.UserCreationRequest;
+import vn.hunghaohan.controller.request.UserPasswordRequest;
+import vn.hunghaohan.controller.request.UserUpdateRequest;
 import vn.hunghaohan.controller.response.UserResponse;
 
 import java.util.List;
 
 public interface UserService {
+
     List<UserResponse> findAll();
+
     UserResponse findById(Long id);
+
     UserResponse findByUsername(String username);
-    int save(UserCreationRequest req);
+
+    Long save(UserCreationRequest req);
+
     int update(UserUpdateRequest req);
+
     void changePassword(UserPasswordRequest req);
+
     void delete(Long id);
 }
